重构代码,全部使用MySQL来保存数据,使用updateTime来代替版本号,直接使用userId分页查询笔记列表

+++ src/main/java/com/cvte/notesync/controller/NoteController.java
@@ -47,7 +47,7 @@ public class NoteController {
 
     @DeleteMapping("/{noteId}/timestamp/{updateTimestamp}")
     @ApiOperation("根据笔记id删除笔记")
-    public Result deleteNode(@PathVariable int noteId, @PathVariable int updateTimestamp) {
+    public Result deleteNode(@PathVariable int noteId, @PathVariable long updateTimestamp) {
         noteService.deleteNode(noteId, updateTimestamp, HolderUtil.getUserId());
         return Result.success();
     }
